% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repipe.R
\name{repipe}
\alias{repipe}
\title{Repipe}
\usage{
repipe()
}
\description{
Replaces magrittr pipes in a highlighted region with native pipes where
possible. Does not replace pipes that are followed by a usage of . to
reference the piped value
}
